requirements.txt
pyautogui
pillow
pytesseract
opencv-python
numpy
pandas
